Adds initial support for skaffold init command and flags.
